Version 2.0
Progress Done:
Converted all plots into functions which are called at the app callback function for dash interractivity.
Used sklearn polynomial features and linear regression to convert regplot to a plot compatible with plotly.
Compleded both tabs of the dashboard.
Added some comments and markdowns.
Issues Oending:
Add line plot of progress in the second tab of the dashboard with data from all players.
Improve code (efficiency and readability).
Add remaning comments and markdowns.

Minor Issues:
Center the checkboxes in the second tab.
Move Gauge charts to the top of the page (one to the left of the current p/l number and one to the right).
Add the players name on top of the first tab with large font.